Responsive/fluid grid via Javascript
Previously, an odd remainder caused a final row that didn't line up
with the pseudo-"columns" above. Now it's left-aligned inline-blocks.
But we still want the whole grid to be centered in its container. And
we want the grid to reflow to fewer columns depending on screen size,
without writing a ton of custom media queries. Maybe that's doable with
nested containers and some mix of text-align:center,
display:inline-block, and margin:0 auto, but I hate that and it
confuses me.